+++ src/components/GridPage/GridPage.jsx
@@ -16,7 +16,8 @@ const GridPage = ({ apiKey }) => {
     setData([])
     axios
       .get(
-        'https://sauti-africa-market-master.herokuapp.com/sauti/developer/filter/',
+        // 'https://sauti-africa-market-master.herokuapp.com/sauti/developer/filter/',
+        'http://localhost:8888/sauti/developer/filter/?currency=USD',
         {
           headers: {
             key: apiKey
